Removed old event loop. Renamed EventLoopV2 -> EventLoop
